Mega commit: task day offsets and task applicability
Applicability is now property of task (previously of "task is in the
checklist" relation). New approach is less flexible, but more simple
to understand and use.

Day offsets tell when the task should be done compared to due date
(starging or leaving). Index page view also shows the due date
of first undone task.